Restructure packaging for upi
COAND-597
