+++ upi/src/main/java/com/adyen/checkout/upi/internal/ui/model/UpiOutputData.kt
@@ -3,16 +3,17 @@
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by oscars on 8/2/2023.
+ * Created by oscars on 21/2/2023.
  */
 
-package com.adyen.checkout.upi
+package com.adyen.checkout.upi.internal.ui.model
 
 import com.adyen.checkout.components.core.internal.ui.model.FieldState
 import com.adyen.checkout.components.core.internal.ui.model.OutputData
 import com.adyen.checkout.components.core.internal.ui.model.Validation
+import com.adyen.checkout.upi.R
 
-class UpiOutputData(
+internal class UpiOutputData(
     val mode: UpiMode,
     virtualPaymentAddress: String,
 ) : OutputData {
